rename AnalyticsContext to Explore
